+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 0.19.0 [2023-11-05]
 
 - Edits a game.
 - Don't verify emails in dev.